feat: reduce to just tests
